Extracted out a bunch of the style structs into separate files.
