+++ components/style/properties/position_longhand.mako.rs
@@ -1,0 +1,70 @@
+<%page args="helpers" />
+
+${helpers.new_style_struct("Position", is_inherited=False, gecko_name="nsStylePosition")}
+
+% for side in ["top", "right", "bottom", "left"]:
+    ${helpers.predefined_type(side, "LengthOrPercentageOrAuto",
+                              "computed::LengthOrPercentageOrAuto::Auto")}
+% endfor
+
+<%helpers:longhand name="z-index">
+    use values::computed::ComputedValueAsSpecified;
+
+    impl ComputedValueAsSpecified for SpecifiedValue {}
+    pub type SpecifiedValue = computed_value::T;
+    pub mod computed_value {
+        use cssparser::ToCss;
+        use std::fmt;
+
+        #[derive(PartialEq, Clone, Eq, Copy, Debug, HeapSizeOf)]
+        pub enum T {
+            Auto,
+            Number(i32),
+        }
+
+        impl ToCss for T {
+            fn to_css<W>(&self, dest: &mut W) -> fmt::Result where W: fmt::Write {
+                match *self {
+                    T::Auto => dest.write_str("auto"),
+                    T::Number(number) => write!(dest, "{}", number),
+                }
+            }
+        }
+
+        impl T {
+            pub fn number_or_zero(self) -> i32 {
+                match self {
+                    T::Auto => 0,
+                    T::Number(value) => value,
+                }
+            }
+        }
+    }
+    #[inline]
+    pub fn get_initial_value() -> computed_value::T {
+        computed_value::T::Auto
+    }
+    fn parse(_context: &ParserContext, input: &mut Parser) -> Result<SpecifiedValue, ()> {
+        if input.try(|input| input.expect_ident_matching("auto")).is_ok() {
+            Ok(computed_value::T::Auto)
+        } else {
+            specified::parse_integer(input).map(computed_value::T::Number)
+        }
+    }
+</%helpers:longhand>
+
+${helpers.predefined_type("min-width", "LengthOrPercentage",
+                          "computed::LengthOrPercentage::Length(Au(0))",
+                          "parse_non_negative")}
+${helpers.predefined_type("max-width", "LengthOrPercentageOrNone",
+                          "computed::LengthOrPercentageOrNone::None",
+                          "parse_non_negative")}
+
+${helpers.predefined_type("min-height", "LengthOrPercentage",
+                          "computed::LengthOrPercentage::Length(Au(0))",
+                          "parse_non_negative")}
+${helpers.predefined_type("max-height", "LengthOrPercentageOrNone",
+                          "computed::LengthOrPercentageOrNone::None",
+                          "parse_non_negative")}
+
+${helpers.single_keyword("box-sizing", "content-box border-box")}
